Comment order comparator by metascore.

+++ src/main/java/com/google/sps/data/Comment.java
@@ -47,8 +47,8 @@ public class Comment {
   public static final Comparator<Comment> ORDER_BY_SCORE = new Comparator<Comment>() {
     @Override
     public int compare(Comment a, Comment b) {
-      return (a.positive.longValue() - a.negative.longValue()) 
-        - (b.positive.longValue() - b.negative.longValue());
+      return (int) ((a.positive.longValue() - a.negative.longValue()) 
+        - (b.positive.longValue() - b.negative.longValue()));
     }
   };
